requirements.txt
openai-whisper
torch
numpy
pandas
pyyaml
soundfile